Change method_id to u64

+++ liteapi/src/tl/utils.rs
@@ -20,9 +20,9 @@ pub fn lossy_read<'tl, T: TlRead<'tl>>(
     }
 }
 
-pub fn method_id(method: &str) -> u32 {
+pub fn method_id(method: &str) -> u64 {
     let crc: u16 = State::<X_25>::calculate(method.as_bytes());
-    (u32::from(crc & 0xffff)) | 0x10000
+    (u64::from(crc & 0xffff)) | 0x10000
 }
 
 pub fn fmt_string(bytes: &[u8], f: &mut std::fmt::Formatter) -> Result<(), std::fmt::Error> {
